Remove old files and rename CMR files in Linodes detail and landing
