+++ packages/manager/src/features/linodes/LinodesDetail/LinodesDetail.tsx
@@ -10,7 +10,6 @@ import { compose } from 'recompose';
 import NotFound from 'src/components/NotFound';
 import SuspenseLoader from 'src/components/SuspenseLoader';
 import useExtendedLinode from 'src/hooks/useExtendedLinode';
-import useFlags from 'src/hooks/useFlags';
 import LinodeDetailErrorBoundary from './LinodeDetailErrorBoundary';
 import {
   LinodeDetailContext,
@@ -19,38 +18,25 @@ import {
 } from './linodeDetailContext';
 
 const LinodesDetailHeader = React.lazy(() => import('./LinodesDetailHeader'));
-const LinodesDetailHeader_CMR = React.lazy(() =>
-  import('./LinodesDetailHeader/LinodeDetailHeader_CMR')
-);
 const LinodesDetailNavigation = React.lazy(() =>
   import('./LinodesDetailNavigation')
 );
-const LinodesDetailNavigation_CMR = React.lazy(() =>
-  import('./LinodesDetailNavigation_CMR')
-);
-const LinodesDashboardNavigation = React.lazy(() =>
-  import('./LinodesDashboardNavigation')
-);
 const CloneLanding = React.lazy(() => import('../CloneLanding'));
-const MigrateLanding = React.lazy(() => import('../MigrateLanding'));
 
 interface Props {
   linodeId: string;
-  isDashboard: boolean;
 }
 
 type CombinedProps = Props & RouteComponentProps<{ linodeId: string }>;
 
 const LinodeDetail: React.FC<CombinedProps> = props => {
   const {
-    isDashboard,
     linodeId,
     match: { path }
   } = props;
 
   const dispatch = useDispatch();
   const linode = useExtendedLinode(+linodeId);
-  const flags = useFlags();
 
   if (!linode) {
     return <NotFound />;
@@ -69,46 +55,14 @@ const LinodeDetail: React.FC<CombinedProps> = props => {
           have to reload all the configs, disks, etc. once we get to the CloneLanding page.
           */}
           <Route path={`${path}/clone`} component={CloneLanding} />
-          {/* With CMR, the Migrate screen no longer has its own route, so some
-          conditional rendering is required here. */}
+
           <Route
-            path={`${path}/migrate`}
-            render={() => {
-              return flags.cmr ? (
-                <React.Fragment>
-                  <LinodesDetailHeader_CMR />
-                  <LinodesDetailNavigation_CMR />
-                </React.Fragment>
-              ) : (
-                <MigrateLanding />
-              );
-            }}
-          />
-          <Route
-            render={() =>
-              flags.cmr ? (
-                // We have separate routing for the version
-                // rendered on the dashboard, to prevent the url
-                // from changing when the active tab is changed.
-                isDashboard ? (
-                  // For single linode view
-                  <React.Fragment>
-                    <LinodesDetailHeader_CMR />
-                    <LinodesDashboardNavigation />
-                  </React.Fragment>
-                ) : (
-                  <React.Fragment>
-                    <LinodesDetailHeader_CMR />
-                    <LinodesDetailNavigation_CMR />
-                  </React.Fragment>
-                )
-              ) : (
-                <React.Fragment>
-                  <LinodesDetailHeader />
-                  <LinodesDetailNavigation />
-                </React.Fragment>
-              )
-            }
+            render={() => (
+              <React.Fragment>
+                <LinodesDetailHeader />
+                <LinodesDetailNavigation />
+              </React.Fragment>
+            )}
           />
         </Switch>
       </React.Suspense>
