+++ packages/manager/src/features/linodes/LinodesDetail/LinodesDetailNavigation.tsx
@@ -2,24 +2,32 @@ import { Config } from '@linode/api-v4/lib/linodes';
 import * as React from 'react';
 import { matchPath, RouteComponentProps, withRouter } from 'react-router-dom';
 import { compose } from 'recompose';
-import SafeTabPanel from 'src/components/SafeTabPanel';
 import TabPanels from 'src/components/core/ReachTabPanels';
 import Tabs from 'src/components/core/ReachTabs';
-import TabLinkList from 'src/components/TabLinkList';
+import { DocumentTitleSegment } from 'src/components/DocumentTitle';
+import SafeTabPanel from 'src/components/SafeTabPanel';
 import SuspenseLoader from 'src/components/SuspenseLoader';
-import VolumesLanding from 'src/features/Volumes/VolumesLanding';
-
+import TabLinkList from 'src/components/TabLinkList';
 import { withLinodeDetailContext } from './linodeDetailContext';
-
-const LinodeSummary = React.lazy(() => import('./LinodeSummary'));
-const LinodeSettings = React.lazy(() => import('./LinodeSettings'));
-const LinodeResize = React.lazy(() => import('./LinodeResize'));
-const LinodeRescue = React.lazy(() => import('./LinodeRescue'));
-const LinodeRebuild = React.lazy(() => import('./LinodeRebuild'));
-const LinodeNetworking = React.lazy(() => import('./LinodeNetworking'));
-const LinodeActivity = React.lazy(() => import('./LinodeActivity'));
-const LinodeAdvanced = React.lazy(() => import('./LinodeAdvanced'));
-const LinodeBackup = React.lazy(() => import('./LinodeBackup'));
+const LinodeSummary_CMR = React.lazy(() =>
+  import('./LinodeSummary/LinodeSummary_CMR')
+);
+const LinodeNetworking_CMR = React.lazy(() =>
+  import('./LinodeNetworking/LinodeNetworking_CMR')
+);
+const LinodeStorage = React.lazy(() => import('./LinodeStorage'));
+const LinodeAdvanced_CMR = React.lazy(() =>
+  import('./LinodeAdvanced/LinodeAdvancedConfigurationsPanel_CMR')
+);
+const LinodeBackup_CMR = React.lazy(() =>
+  import('./LinodeBackup/LinodeBackup_CMR')
+);
+const LinodeActivity_CMR = React.lazy(() =>
+  import('./LinodeActivity/LinodeActivity_CMR')
+);
+const LinodeSettings_CMR = React.lazy(() =>
+  import('./LinodeSettings/LinodeSettings_CMR')
+);
 
 type CombinedProps = ContextProps &
   RouteComponentProps<{
@@ -28,39 +36,26 @@ type CombinedProps = ContextProps &
 
 const LinodesDetailNavigation: React.FC<CombinedProps> = props => {
   const {
-    match: { url },
     linodeLabel,
-    linodeConfigs,
-    linodeId,
-    linodeRegion,
-    readOnly
+    match: { url }
   } = props;
 
   const tabs = [
-    /* NB: These must correspond to the routes inside the Switch */
     {
-      routeName: `${url}/summary`,
-      title: 'Summary'
-    },
-    {
-      routeName: `${url}/volumes`,
-      title: 'Volumes'
+      routeName: `${url}/analytics`,
+      title: 'Analytics'
     },
     {
       routeName: `${url}/networking`,
-      title: 'Networking'
+      title: 'Network'
     },
     {
-      routeName: `${url}/resize`,
-      title: 'Resize'
+      routeName: `${url}/storage`,
+      title: 'Storage'
     },
     {
-      routeName: `${url}/rescue`,
-      title: 'Rescue'
-    },
-    {
-      routeName: `${url}/rebuild`,
-      title: 'Rebuild'
+      routeName: `${url}/configurations`,
+      title: 'Configurations'
     },
     {
       routeName: `${url}/backup`,
@@ -68,15 +63,11 @@ const LinodesDetailNavigation: React.FC<CombinedProps> = props => {
     },
     {
       routeName: `${url}/activity`,
-      title: 'Activity'
+      title: 'Activity Feed'
     },
     {
       routeName: `${url}/settings`,
       title: 'Settings'
-    },
-    {
-      routeName: `${url}/advanced`,
-      title: 'Disks/Configs'
     }
   ];
 
@@ -84,71 +75,60 @@ const LinodesDetailNavigation: React.FC<CombinedProps> = props => {
     return Boolean(matchPath(p, { path: location.pathname }));
   };
 
+  const getIndex = () => {
+    return Math.max(
+      tabs.findIndex(tab => matches(tab.routeName)),
+      0
+    );
+  };
+
   const navToURL = (index: number) => {
     props.history.push(tabs[index].routeName);
   };
 
   return (
-    <Tabs
-      index={Math.max(
-        tabs.findIndex(tab => matches(tab.routeName)),
-        0
-      )}
-      onChange={navToURL}
-    >
-      <TabLinkList tabs={tabs} />
+    <>
+      <DocumentTitleSegment
+        segment={`${linodeLabel} - ${tabs[getIndex()]?.title ?? 'Detail View'}`}
+      />
+      <div style={{ marginTop: 8 }}>
+        <Tabs index={getIndex()} onChange={navToURL}>
+          <TabLinkList tabs={tabs} />
 
-      <React.Suspense fallback={<SuspenseLoader />}>
-        <TabPanels>
-          <SafeTabPanel index={0}>
-            <LinodeSummary />
-          </SafeTabPanel>
-          <SafeTabPanel index={1}>
-            <VolumesLanding
-              linodeId={linodeId}
-              linodeLabel={linodeLabel}
-              linodeRegion={linodeRegion}
-              linodeConfigs={linodeConfigs}
-              readOnly={readOnly}
-              fromLinodes
-              removeBreadCrumb
-            />
-          </SafeTabPanel>
+          <React.Suspense fallback={<SuspenseLoader />}>
+            <TabPanels>
+              <SafeTabPanel index={0}>
+                <LinodeSummary_CMR />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={2}>
-            <LinodeNetworking />
-          </SafeTabPanel>
+              <SafeTabPanel index={1}>
+                <LinodeNetworking_CMR />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={3}>
-            <LinodeResize />
-          </SafeTabPanel>
+              <SafeTabPanel index={2}>
+                <LinodeStorage />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={4}>
-            <LinodeRescue />
-          </SafeTabPanel>
+              <SafeTabPanel index={3}>
+                <LinodeAdvanced_CMR />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={5}>
-            <LinodeRebuild />
-          </SafeTabPanel>
+              <SafeTabPanel index={4}>
+                <LinodeBackup_CMR />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={6}>
-            <LinodeBackup />
-          </SafeTabPanel>
+              <SafeTabPanel index={5}>
+                <LinodeActivity_CMR />
+              </SafeTabPanel>
 
-          <SafeTabPanel index={7}>
-            <LinodeActivity />
-          </SafeTabPanel>
-
-          <SafeTabPanel index={8}>
-            <LinodeSettings />
-          </SafeTabPanel>
-
-          <SafeTabPanel index={9}>
-            <LinodeAdvanced />
-          </SafeTabPanel>
-        </TabPanels>
-      </React.Suspense>
-    </Tabs>
+              <SafeTabPanel index={6}>
+                <LinodeSettings_CMR />
+              </SafeTabPanel>
+            </TabPanels>
+          </React.Suspense>
+        </Tabs>
+      </div>
+    </>
   );
 };
